JDP190401-13 Group Entity implementation
