+++ src/main/java/com/kodilla/ecommercee/dto/Group_Entity.java
@@ -1,0 +1,31 @@
+package com.kodilla.ecommercee.dto;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+
+@Entity
+public class Group_Entity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String description;
+
+    public Group_Entity() {
+    }
+
+    public Group_Entity(String description) {
+        this.description = description;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
